Fix worker isolation
When using a database pool, as we did in the code we stripped down,
there you have the pool, get a connection, and then create a prepared
statement. I've simplified this to just getting a "Worker" from the
Service, and use it. Of course, I now need to get "another" Worker each
time I get one, since we don't really have a pool and a prepared
statement anymore. So I've done that, and it should fix any potential
race conditions that would occur otherwise at the second batch.

Signed-off-by: Matthias Nott <matthias.nott@sap.com>